requirements.txt
torch 
torchvision 
torchaudio --index-url https://download.pytorch.org/whl/cu121
opencv-python
PyYAML
tqdm
thop
matplotlib
pymongo 
pillow 
pyyaml